+++ src/route.h
@@ -2,16 +2,17 @@
 #ifndef ROUTE_H
 #define ROUTE_H
 
+#include <cstdint>
 #include <fstream>
 #include <vector>
 
 class Route {
 private:
-  std::vector<int> path;
+  std::vector<int64_t> path;
 
 public:
-  Route(std::vector<int> path);
-  int getPosition(int index);
+  Route(std::vector<int64_t> path);
+  int64_t getPosition(int index);
   int getLength();
   static Route fromFile(std::ifstream &file);
 };
